+++ client/src/app/budget/budgets.tsx
@@ -16,46 +16,7 @@ import { type AxiosResponse } from 'axios';
 import { defaultGuid } from '@/types/user';
 import ResponsiveButton from '@/components/responsive-button';
 import AlertBanner from '@/components/alert-banner';
-import UnbudgetCard from './unbudget-card';
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger,
-} from '@/components/ui/accordion';
-
-interface Unbudget {
-  category: string;
-  amount: number;
-}
-
-const getUnbudgetedTransactions = (budgets: Budget[], transactions: Transaction[]): Unbudget[] => {
-  if (budgets == null || transactions == null) return [];
-
-  // This creates an object that maps category => array of amounts
-  const groupedTransactions: [string, number[]][] = Object.entries(
-    transactions.reduce((result: any, item: Transaction) => {
-      (result[item['category']] = result[item['category']] || []).push(item.amount);
-      return result;
-    }, {})
-  );
-
-  const filteredGroupedTransactions = groupedTransactions.filter((t) => {
-    return !budgets.some(({ category }) => category === t[0]);
-  });
-
-  const unbudgetedTransactions: Unbudget[] = [];
-  filteredGroupedTransactions.forEach((element) => {
-    unbudgetedTransactions.push({
-      category: element[0],
-      amount: element[1].reduce((a, b) => {
-        return a + b;
-      }),
-    });
-  });
-
-  return unbudgetedTransactions;
-};
+import Unbudgets from './unbudgets';
 
 const Budgets = (): JSX.Element => {
   const [date, setDate] = React.useState<Date>(initMonth());
@@ -164,45 +125,15 @@ const Budgets = (): JSX.Element => {
               )}
               isPending={budgetsQuery.isPending || transactionsQuery.isPending}
             />
-            <Accordion type="single" collapsible className="w-2/3">
-              <AccordionItem value="item-1">
-                <AccordionTrigger>
-                  <div className="text-l scroll-m-20 justify-self-start font-semibold tracking-tight">
-                    Unbudgeted Transactions
-                  </div>
-                  <div className="text-l scroll-m-20 justify-self-start font-semibold tracking-tight">
-                    {getUnbudgetedTransactions(
-                      budgetsQuery.data?.data,
-                      getTransactionsForMonth(
-                        (transactionsQuery.data?.data as Transaction[]) ?? [],
-                        date
-                      )
-                    )
-                      .reduce((a: number, b: Unbudget) => {
-                        return a + b['amount'];
-                      }, 0)
-                      .toFixed()}
-                  </div>
-                </AccordionTrigger>
-                <AccordionContent>
-                  {getUnbudgetedTransactions(
-                    budgetsQuery.data?.data,
-                    getTransactionsForMonth(
-                      (transactionsQuery.data?.data as Transaction[]) ?? [],
-                      date
-                    )
-                  ).map((unbudget: Unbudget) => (
-                    <UnbudgetCard
-                      key={unbudget.category}
-                      name={unbudget.category}
-                      amount={unbudget.amount.toFixed()}
-                    />
-                  ))}
-                </AccordionContent>
-              </AccordionItem>
-            </Accordion>
           </div>
         </div>
+        <Unbudgets
+          transactions={getTransactionsForMonth(
+            (transactionsQuery.data?.data as Transaction[]) ?? [],
+            date
+          )}
+          budgets={(budgetsQuery.data?.data as Budget[]) ?? []}
+        />
       </div>
       <div className="flex h-96 w-1/4 flex-col justify-center">
         {/* TODO: Figure out a better way to horizontally position this */}
